+++ server-spi-private/src/main/java/org/keycloak/models/utils/KeycloakModelUtils.java
@@ -497,16 +497,18 @@ public final class KeycloakModelUtils {
 
     }
 
-    public static List<String>  resolveAttribute(GroupModel group, String name) {
-        List<String> values = group.getAttributeStream(name).collect(Collectors.toList());
+    public static Set<String>  resolveAttribute(GroupModel group, List<String> names) {
+        Set<String> values = new HashSet<>();
+        values.addAll(names.stream().flatMap(name -> group.getAttributeStream(name)).collect(Collectors.toList()));
         if (!values.isEmpty()) return values;
         if (group.getParentId() == null) return null;
-        return resolveAttribute(group.getParent(), name);
+        return resolveAttribute(group.getParent(), names);
     }
 
 
-    public static Collection<String> resolveAttribute(UserModel user, String name, boolean aggregateAttrs) {
-        List<String> values = user.getAttributeStream(name).collect(Collectors.toList());
+    public static Collection<String> resolveAttribute(UserModel user, List<String> names, boolean aggregateAttrs) {
+        Set<String> values = new HashSet<>();
+        values.addAll(names.stream().flatMap(name -> user.getAttributeStream(name)).collect(Collectors.toList()));
         Set<String> aggrValues = new HashSet<String>();
         if (!values.isEmpty()) {
             if (!aggregateAttrs) {
@@ -514,13 +516,13 @@ public final class KeycloakModelUtils {
             }
             aggrValues.addAll(values);
         }
-        Stream<List<String>> attributes = user.getGroupsStream()
-                .map(group -> resolveAttribute(group, name))
+        Stream<Set<String>> attributes = user.getGroupsStream()
+                .map(group -> resolveAttribute(group, names))
                 .filter(Objects::nonNull)
                 .filter(attr -> !attr.isEmpty());
 
         if (!aggregateAttrs) {
-            Optional<List<String>> first = attributes.findFirst();
+            Optional<Set<String>> first = attributes.findFirst();
             if (first.isPresent()) return first.get();
         } else {
             aggrValues.addAll(attributes.flatMap(Collection::stream).collect(Collectors.toSet()));